Move remaining classes from ItemLoader to ItemSearch
