+++ src/Services/ItemSearch/Contracts/FacetExtension.php
@@ -1,5 +1,5 @@
 <?php
-namespace IO\Services\ItemLoader\Contracts;
+namespace IO\Services\ItemSearch\Contracts;
 
 use Plenty\Modules\Cloud\ElasticSearch\Lib\Search\Aggregation\AggregationInterface;
