+++ src/Services/ItemSearch/Factories/VariationSearchFactory.php
@@ -4,8 +4,7 @@ namespace IO\Services\ItemSearch\Factories;
 
 use IO\Helper\CurrencyConverter;
 use IO\Helper\VatConverter;
-use IO\Services\ItemLoader\Contracts\FacetExtension;
-use IO\Services\ItemLoader\Services\FacetExtensionContainer;
+use IO\Services\ItemSearch\Contracts\FacetExtension;
 use IO\Services\ItemSearch\Extensions\GroupedAttributeValuesExtension;
 use IO\Services\ItemSearch\Extensions\BundleComponentExtension;
 use IO\Services\ItemSearch\Extensions\ContentCacheVariationLinkExtension;
@@ -14,6 +13,7 @@ use IO\Services\ItemSearch\Extensions\ItemDefaultImage;
 use IO\Services\ItemSearch\Extensions\ItemUrlExtension;
 use IO\Services\ItemSearch\Extensions\PriceSearchExtension;
 use IO\Services\ItemSearch\Extensions\ReduceDataExtension;
+use IO\Services\ItemSearch\Helper\FacetExtensionContainer;
 use IO\Services\ItemSearch\Mutators\OrderPropertySelectionValueMutator;
 use IO\Services\PriceDetectService;
 use IO\Services\SessionStorageService;
